refactor: correct the import path of `cn` util function
